特性：重构性能监控为 RAM 监控与应用排行优化
本次提交主要对“性能监控”卡片进行了维度调整，由 CPU 监控变更为内存（RAM）监控，并对应用使用排行卡片进行了数据类型升级与 UI 布局优化。

**性能监控卡片重构 (`CardPerformance`)**

*   **监控指标变更**:
    *   移除 `CpuMonitor` 相关逻辑，改为监控系统 **RAM 使用情况**。
    *   新增 `RamInfo` 实体类，通过 `ActivityManager` 获取总内存、已用内存及可用内存。
*   **UI 全新设计**:
    *   修改 `fragment_card_performance.xml`，引入环形进度条 (`pb_ram`) 展示内存占用百分比。
    *   新增具体数值文本，实时显示“已用 / 总共”内存大小（如 "64 GB / 128 GB"）。
    *   保留 `WaveformView` 并调整其布局比例，用于展示内存占用的动态波形。

**波形图组件优化 (`WaveformView`)**

*   **动态自适应**:
    *   新增自动缩放逻辑，引入 `maxValues` 和 `MAX_DECAY`，使波形图能根据输入数据的幅值动态调整纵轴比例。
    *   增加 `HEADROOM` (顶部留白)，避免波形顶格绘制。

**应用使用排行优化**

*   **数据结构升级**:
    *   将 `AppInfo` 及数据库中的 `openAppCount`（启动次数）字段类型由 `int` 更改为 `long`，以支持更大数值。
    *   在 `AllAppsInfoDao` 中将排行查询限制为前 3 名 (`LIMIT 3`)。
*   **列表项 UI 调整**:
    *   重构 `item_app_rank.xml`，将计数文本 (`tv_count`) 移至应用名称下方，并缩小字号。
    *   `AppUsageRankAdapter` 新增数值格式化逻辑，支持 "k" (千) 和 "M" (百万) 单位显示。

**通用变更与资源清理**

*   **进度条着色逻辑**:
    *   在 `BaseCardFragment` 中统一实现了 `getProgressDrawable` 方法，根据百分比返回颜色值（绿/橙/红）。
    *   `CardStorage` 和 `CardPerformance` 改用 `setProgressTintList` 进行动态着色，不再通过替换 Drawable 资源实现。
*   **资源清理**:
    *   删除了不再使用的 `bg_storage_progress_ring_green.xml`、`_orange.xml` 和 `_red.xml` 文件。

+++ app/src/main/java/com/jayjd/boxtop/entity/AppInfo.java
@@ -30,7 +30,7 @@ public class AppInfo implements Serializable {
 
     private int isBanner;
 
-    private int openAppCount;
+    private long openAppCount;
     private int sortIndex;
     private int cardColor;
     private String packagePath;
